TST: fix tests
